Add `modify-interactive` command
This command is intended to be used for classifying remaining devices that cannot be easily classified via regex due to bad/illogical device naming schemes (i'm looking at you China).

The idea is that a Google image search will be performed for each device that is unclassified, prompting for a classification and then moving onto the next.